Document adding classes Serialized by LabTech to Psych
We've been using LabTech for about a week at one of my clients, and we
had a... tiny bit of an outage because our test-cases didn't quite cover
running the experiment when it was *deserializing* data.

We're not sure if this is a comprehensive list of classes used by
LabTech, because we had some classes already permitted
(`ActiveSupport::Duration`, `Range`, and
`ActiveSupport::HashWithIndifferentAccess` to be specific).

Hopefully this will help other folks getting started to avoid any
unpleasantness.